Auth on recipes, adding ingredient repo

+++ server/Services/RecipeService.cs
@@ -6,13 +6,15 @@ public class RecipeService(RecipeRepo recipeRepo)
   { return recipeRepo.GetRecipes(); }
 
   internal Recipe GetRecipeById(int recipeId)
-  { return recipeRepo.GetRecipeById(recipeId); }
+  { return recipeRepo.GetRecipeById(recipeId) ?? throw new Exception($"Recipe with ID {recipeId} does not exist"); }
 
   internal Recipe CreateRecipe(Recipe recipeData)
   { return recipeRepo.CreateRecipe(recipeData); }
 
-  internal string DeleteRecipe(int recipeId)
+  internal string DeleteRecipe(string creatorId, int recipeId)
   {
+    Recipe recipe = GetRecipeById(recipeId);
+    if (recipe.CreatorId != creatorId) { throw new Exception("Forbidden action - Not your data"); }
     recipeRepo.DeleteRecipe(recipeId);
     return "Recipe deleted";
   }
